requirements.txt
huggingface_hub>=0.20
openai>=1.40
python-dotenv>=1.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.13
